restructure of program with classes and member functions
