+++ database.hpp
@@ -1,0 +1,72 @@
+#include <iostream>
+#include <vector>
+#include <fstream>
+#include <sstream>
+#include <string>
+#include <map> //linking documents with databases
+#include <algorithm>
+//#include <nlohmann/json.hpp>
+
+using namespace std;
+
+//chatgpt generated sample document structure to store document in database
+struct Movie_Document {
+    string poster_Link;
+    string series_title;
+    int released_year;
+    string certificate;
+    int runtime; //in minutes
+    string genre;
+    double IMDB_rating; //specific for IMDB movie data, scale 10
+    string overview;
+    int meta_score; //scale 100
+    string Director;
+    string Star1;
+    string Star2;
+    string Star3;
+    string Star4;
+    int numVotes;
+    int gross;
+    //vector <string> stars; //keep all stars
+};
+
+//simulating MongoDB's feature of changing database
+struct collection {
+    string name;
+    vector<Movie_Document*> getMovieDocs(){return this->movieDocs;}
+    vector <Movie_Document*> movieDocs;  //documents linked with this DataBase
+};
+
+class Database{
+
+    public:
+        Database();
+        void dbAll();
+        void addCollection(string cltName);
+        void useCollection(string cltName);
+        void printCurrentClt();
+        //collection* useCollection(string cltName);
+        void importCSV(string cltName, string fileName);
+        void importJSON(collection* db, string fileName);
+        void exportData(collection* cc, Movie_Document* movDoc, string format);
+        void addDocumentManually(collection*);
+        void addDocManual(string cltName, Movie_Document* movDoc);
+        void deleteCollections(string cltName);
+        void deleteAllDocs(string cltName);
+        //Movie_Document* updateEntry(string cltName);
+        void deleteDocManual();
+        collection* getCollectionByName(string name);
+        bool deleteCollectionByName(string name);
+        void print();
+        void printAllClt();
+        void printSingleClt(string name);
+        void createCollection(string name);
+        void updateEntry(string cltName);
+        vector<string> getCollectionsList();
+
+
+    private:
+        collection* currentClt;
+        vector<collection*> collectionDB;
+
+};
